feat: backend-lint-package-size-004 - Bring checked backend packages under the cap
