+++ docs/internal/processes/development-guide-relevant-files.md
@@ -161,7 +161,8 @@ This inventory records the checked-in files and directories that the maintainer 
 | `pkg/factory/projections/simple_dashboard_projection.go` | Simple-dashboard runtime projection owner | Customer-visible runtime counts and place/work-item maps for the lightweight dashboard seam should stay derived here from reconstructed world state so active, non-terminal, terminal, and failed transition views remain aligned across backend projection consumers. |
 | `pkg/config/layout.go` | Canonical factory clone owner | Root-level `FactoryConfig` fields such as authored `factory.guards` must be copied through `CloneFactoryConfig` so `LoadRuntimeConfig` and other layout flows do not silently drop supported customer policy; hosted worker `Auth` and `Linear` deep-copy policy should stay delegated to the canonical same-package helpers in `pkg/config/agents_config.go` instead of being recopied inline here; portable expand and named-factory persistence should also keep thin imported `factory.json` generation behind `authoredFactoryConfigForExpandedLayout(...)` before flattening so regression tests assert the same authored-layout contract users actually get on disk. |
 | `pkg/config/layouttests/layout_clone_test.go` | Layout clone behavioral regression surface | `CloneWorkerConfig(...)` and `CloneFactoryConfig(...)` regression coverage should assert observable detached-copy behavior from an external test package here by mutating the source config after cloning and checking the cloned outcome, rather than asserting helper placement, helper names, or source-file inventories. |
-| `pkg/config/factory_config_mapping_test.go` | Generated factory-config boundary regression surface | Boundary rejection regressions should stay focused here by asserting `generatedFactoryBoundaryErrorPrefix` plus the full path-specific retirement guidance string for top-level, nested `definition`, and nested `definition.cron` aliases. |
+| `pkg/config/maptests/`, `pkg/config/mappingtests/`, `pkg/config/runtimetests/`, and `pkg/config/exhaustiontests/` | Config package-size test split seam | Keep exported-surface mapper, generated-boundary mapping, runtime-load, named-factory, and contract-guard coverage in focused external test packages when root `pkg/config` approaches the file-count cap. Keep root config tests only for package-private semantic validator, layout materialization, and staged persistence helpers that should not be exported for test movement. |
+| `pkg/config/mappingtests/` | Generated factory-config boundary regression surface | Boundary rejection regressions should stay focused here by asserting the generated-boundary error prefix plus the full path-specific retirement guidance string for top-level, nested `definition`, and nested `definition.cron` aliases. |
 | `pkg/config/openapitests/` | Generated OpenAPI JSON boundary regression surface | When cleanup work must preserve generated-boundary error strings or prove the generated `Factory` JSON shape itself, keep black-box proof here with `GeneratedFactoryFromOpenAPIJSON(...)` assertions over the generated-boundary error prefix, exact field paths, replacement guidance, and marshaled `factoryapi.Factory` workstation keys. |
 | `pkg/replay/event_stream_artifact.go` | Replay event-stream import and adjacent-factory reconstruction owner | SSE replay import, truncated-tail recovery, legacy cron compatibility normalization, and adjacent authored-factory hydration plus payload rewriting should stay centralized here so saved event streams rebuild the same backend-visible factory shape as canonical replay artifacts; tests that need file-based conversion should call `ArtifactFromEventStreamFile` or `SaveArtifactFromEventStreamFile` instead of copying that hydration path into suite-local helpers. |
 | `pkg/config/portable_bundled_files.go` | Supported portable bundled-file collection and materialization seam | Export/read-back portability changes should extend this seam so disk-backed `scripts/`, `docs/`, and allowlisted helper files stay mapped onto the canonical `factory/...` bundled-file contract for portable `factory/` layouts and named persisted factories without broadening generic runtime copy semantics; thin authored manifests should rehydrate existing supported entries from disk here before any broader default collection runs, portable script entries that omit inline content still need load-time mode normalization because they bypass the write path that already calls `os.Chmod(...)`, and overwrite detection for differing bundled-file materialization should stay centralized here so higher layers consume one replacement report instead of re-diffing on their own through `ExpandFactoryConfigLayoutWithReport(...)` or `LoadedFactoryConfig.PortableBundledFileReplacements()`. Maintainability cleanup in this seam should keep entry filtering, relative-path validation, and file-materialization helpers local so portable path policy and observable filesystem errors stay shared across export, expand, and starter-work collection flows. |
