+++ pkg/config/mappingtests/factory_config_mapping_roundtrip_test.go
@@ -1,7 +1,8 @@
-package config
+package mappingtests
 
 import (
 	"encoding/json"
+	. "github.com/portpowered/infinite-you/pkg/config"
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 	"strings"
 	"testing"
