+++ pkg/config/mappingtests/helpers_test.go
@@ -1,0 +1,7 @@
+package mappingtests
+
+const generatedFactoryBoundaryErrorPrefix = "decode factory generated-schema boundary"
+
+func stringPtr(value string) *string {
+	return &value
+}
