+++ pkg/config/maptests/config_mapper_equivalence_test.go
@@ -1,8 +1,9 @@
-package config
+package maptests
 
 import (
 	"context"
 	"fmt"
+	. "github.com/portpowered/infinite-you/pkg/config"
 	"strings"
 	"testing"
