+++ pkg/config/runtimetests/runtime_config_named_factory_test.go
@@ -1,7 +1,8 @@
-package config
+package runtimetests
 
 import (
 	"encoding/json"
+	. "github.com/portpowered/infinite-you/pkg/config"
 	"os"
 	"path/filepath"
 	"strings"
@@ -201,45 +202,6 @@ func TestPersistNamedFactory_RejectsInvalidPayloadWithoutChangingCurrentFactory(
 	}
 	if loaded.FactoryConfig().Project != "alpha" {
 		t.Fatalf("project after invalid persist = %q, want alpha", loaded.FactoryConfig().Project)
-	}
-}
-
-func TestPersistNamedFactory_RollsBackStagedLayoutWhenLoadRuntimeConfigFails(t *testing.T) {
-	rootDir := t.TempDir()
-
-	if _, err := PersistNamedFactory(rootDir, "alpha", namedFactoryPayload(t, "alpha")); err != nil {
-		t.Fatalf("PersistNamedFactory(alpha): %v", err)
-	}
-	if err := WriteCurrentFactoryPointer(rootDir, "alpha"); err != nil {
-		t.Fatalf("WriteCurrentFactoryPointer(alpha): %v", err)
-	}
-
-	_, err := persistNamedFactory(rootDir, "broken", namedFactoryPayload(t, "broken"), namedFactoryPersistOptions{}, namedFactoryPersistHooks{
-		afterWrite: func(stagingDir string) error {
-			path := filepath.Join(stagingDir, interfaces.WorkstationsDir, "execute-broken", interfaces.FactoryAgentsFileName)
-			return os.WriteFile(path, []byte("---\ntype: [\n"), 0o644)
-		},
-	})
-	if err == nil {
-		t.Fatal("expected staged named-factory validation failure")
-	}
-	if got := err.Error(); !containsAll(got, `validate factory "broken" config`, "load workstation", "AGENTS.md missing closing frontmatter delimiter") {
-		t.Fatalf("expected load-time validation error, got %v", err)
-	}
-
-	if _, err := os.Stat(filepath.Join(rootDir, "broken")); !os.IsNotExist(err) {
-		t.Fatalf("expected failed staged factory directory to be absent, got stat err=%v", err)
-	}
-
-	loaded, err := LoadRuntimeConfig(rootDir, nil)
-	if err != nil {
-		t.Fatalf("LoadRuntimeConfig(current after staged failure): %v", err)
-	}
-	if loaded.FactoryDir() != filepath.Join(rootDir, "alpha") {
-		t.Fatalf("FactoryDir after staged failure = %q, want %q", loaded.FactoryDir(), filepath.Join(rootDir, "alpha"))
-	}
-	if loaded.FactoryConfig().Project != "alpha" {
-		t.Fatalf("project after staged failure = %q, want alpha", loaded.FactoryConfig().Project)
 	}
 }
